Refactors authentication to use username
Updates the authentication process to use username instead of email for login.
This change improves user experience by allowing users to login with their
preferred username. It also updates the discord registration process to
collect a username and store it in the database, as well as updating the docker
image to automatically build and deploy on push to main.